code_companion/requirements.txt
qdrant-client 
openai 
python-dotenv 
tiktoken
transformers